% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeatmapWidget.R
\name{HeatmapWidget}
\alias{HeatmapWidget}
\alias{printf}
\title{HeatmapWidget}
\usage{
printf(...)
}
\value{
shiny code which, wnen invoked (almost always by the shinyApp function, returns html

nothing
}
\description{
an R6 class providing clean access to iheatmapr
}
\details{
R6 Class representing a simple text display widget
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{HeatmapWidget$new()}}
\item \href{#method-ui}{\code{HeatmapWidget$ui()}}
\item \href{#method-server}{\code{HeatmapWidget$server()}}
\item \href{#method-setHeatmap}{\code{HeatmapWidget$setHeatmap()}}
\item \href{#method-clone}{\code{HeatmapWidget$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new HeatmapWidget
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HeatmapWidget$new(
  id,
  title,
  mtx,
  rowTitle,
  columnTitle,
  clusteringMethod = "hclust",
  rowClusters = 5,
  colClusters = 5,
  width = 600,
  height = 600,
  quiet = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{the html document div id}

\item{\code{title}}{character}

\item{\code{mtx}}{a numeric matrix}

\item{\code{rowTitle}}{character, displayed on far left of the plot}

\item{\code{columnTitle}}{character, displayed at the bottom}

\item{\code{clusteringMethod}}{character, "hclust" (the default), or "kmeans"}

\item{\code{rowClusters}}{integer, the number of k-means row clusters desired, default 5}

\item{\code{colClusters}}{integer, the number of k-means column clusters desired, default 5}

\item{\code{width}}{numeric, pixels, default 600}

\item{\code{height}}{numeric, pixels, default 600}

\item{\code{quiet}}{logical silent or verbose?  default TRUE}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `HeatmapWidget` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ui"></a>}}
\if{latex}{\out{\hypertarget{method-ui}{}}}
\subsection{Method \code{ui()}}{
defines the html structure of this widget
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HeatmapWidget$ui()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-server"></a>}}
\if{latex}{\out{\hypertarget{method-server}{}}}
\subsection{Method \code{server()}}{
defines the event handlers used with the ui
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HeatmapWidget$server(input, output, session)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{list, managed by Shiny}

\item{\code{output}}{list, managed by Shiny}

\item{\code{session}}{list (environment?) managed by Shiny}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setHeatmap"></a>}}
\if{latex}{\out{\hypertarget{method-setHeatmap}{}}}
\subsection{Method \code{setHeatmap()}}{
new heatmap displayed
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HeatmapWidget$setHeatmap(
  mtx,
  clusteringMethod = "hclust",
  rowClusters = 5,
  colClusters = 5
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mtx}}{numeric matrix}

\item{\code{clusteringMethod}}{character, "hclust" by default}

\item{\code{rowClusters}}{numeric, 5 by default}

\item{\code{colClusters}}{numeric, 5 by default}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HeatmapWidget$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
